+++ VMagicMirrorConfig/VMagicMirrorConfig/ViewModel/BlendShapeItemViewModel.cs
@@ -1,6 +1,4 @@
-﻿using System.Xml.Serialization;
-
-namespace Baku.VMagicMirrorConfig
+﻿namespace Baku.VMagicMirrorConfig
 {
     public class BlendShapeItemViewModel : ViewModelBase
     {
@@ -14,10 +12,10 @@ namespace Baku.VMagicMirrorConfig
         /// <param name="isUsedWithThisAvatar"></param>
         public BlendShapeItemViewModel(WordToMotionItemViewModel parent, string name, int value, bool isUsedWithThisAvatar)
         {
-            _parent = parent;
             BlendShapeName = name;
             ValuePercentage = value;
             IsUsedWithThisAvatar = isUsedWithThisAvatar;
+            ForgetThisClipCommand = new ActionCommand(() => parent.RequestForgetClip(this));
         }
 
         /// <summary>
@@ -30,8 +28,6 @@ namespace Baku.VMagicMirrorConfig
             this(parent, name, value, true)
         {
         }
-
-        private readonly WordToMotionItemViewModel _parent;
 
         public string BlendShapeName { get; }
 
@@ -49,17 +45,12 @@ namespace Baku.VMagicMirrorConfig
         /// また基本以外のブレンドシェイプクリップについては、
         /// アバターが一度もロードされていない場合はtrueでよいものとします。
         /// </summary>
-        [XmlIgnore]
         public bool IsUsedWithThisAvatar
         { 
             get => _isUsedWithThisAvatar;
             set => SetValue(ref _isUsedWithThisAvatar, value);
         }
 
-        private ActionCommand? _forgetThisClipCommand = null;
-        public ActionCommand ForgetThisClipCommand
-            => _forgetThisClipCommand ??= new ActionCommand(RequestForgetThisClip);
-        private void RequestForgetThisClip()
-            => _parent.RequestForgetClip(this);
+        public ActionCommand ForgetThisClipCommand { get; }
     }
 }
