+++ VMagicMirrorConfig/VMagicMirrorConfig/ViewModel/GamepadSettingViewModel.cs
@@ -1,5 +1,6 @@
 ﻿namespace Baku.VMagicMirrorConfig
 {
+    /// <summary>ゲームパッド関係の設定のViewModel。素通し設定ばかりなのでとても単純。</summary>
     public class GamepadSettingViewModel : SettingViewModelBase
     {
         internal GamepadSettingViewModel(GamepadSettingModel model, IMessageSender sender) : base(sender)
